requirements.txt
flask==2.3.3
werkzeug==2.3.7
gunicorn==21.2.0
sqlitecloud==1.3.0
python-dotenv==1.0.0